Changed from float to double

+++ src/main/java/Division.java
@@ -2,15 +2,15 @@
  * Simple class to do a division between two numbers
  */
 public class Division {
-    private float firstNumber;
-    private float secondNumber;
+    private double firstNumber;
+    private double secondNumber;
 
     /**
      *
      * @param firstNumber the dividend of the division
      * @param secondNumber the divisor of the division
      */
-    public Division(float firstNumber, float secondNumber) {
+    public Division(double firstNumber, double secondNumber) {
         this.firstNumber = firstNumber;
         this.secondNumber = secondNumber;
     }
@@ -19,7 +19,7 @@ public class Division {
      *
      * @return the quotient of the division
      */
-    public float doDivision() {
+    public double doDivision() {
         if (secondNumber != 0) {
             return firstNumber / secondNumber;
         } else {
